+++ jib-core/src/main/java/com/google/cloud/tools/jib/frontend/JavaLayerConfigurations.java
@@ -36,33 +36,29 @@ public class JavaLayerConfigurations {
   /** Represents the different types of layers for a Java application. */
   @VisibleForTesting
   enum LayerType {
-    DEPENDENCIES("dependencies", JavaEntrypointConstructor.DEFAULT_DEPENDENCIES_PATH_ON_IMAGE),
-    SNAPSHOT_DEPENDENCIES(
-        "snapshot dependencies", JavaEntrypointConstructor.DEFAULT_DEPENDENCIES_PATH_ON_IMAGE),
-    RESOURCES("resources", JavaEntrypointConstructor.DEFAULT_RESOURCES_PATH_ON_IMAGE),
-    CLASSES("classes", JavaEntrypointConstructor.DEFAULT_CLASSES_PATH_ON_IMAGE),
+    DEPENDENCIES("dependencies"),
+    SNAPSHOT_DEPENDENCIES("snapshot dependencies"),
+    RESOURCES("resources"),
+    CLASSES("classes"),
     // TODO: remove this once we put files in WAR into the relevant layers (i.e., dependencies,
     // snapshot dependencies, resources, and classes layers). Should copy files in the right
-    EXPLODED_WAR("exploded war", JavaEntrypointConstructor.DEFAULT_JETTY_BASE_ON_IMAGE),
-    EXTRA_FILES("extra files", "/");
+    EXPLODED_WAR("exploded war"),
+    EXTRA_FILES("extra files");
 
     private final String name;
-    private final Path extractionPath;
 
-    /** Initializes with a name for the layer and the layer files' default extraction path root. */
-    LayerType(String name, String extractionPath) {
+    /**
+     * Initializes with a name for the layer.
+     *
+     * @param name name to set for the layer; does not affect the contents of the layer
+     */
+    LayerType(String name) {
       this.name = name;
-      this.extractionPath = Paths.get(extractionPath);
     }
 
     @VisibleForTesting
     String getName() {
       return name;
-    }
-
-    @VisibleForTesting
-    Path getExtractionPath() {
-      return extractionPath;
     }
   }
 
@@ -70,42 +66,51 @@ public class JavaLayerConfigurations {
   public static class Builder {
 
     private final Map<LayerType, List<Path>> layerFilesMap = new EnumMap<>(LayerType.class);
+    private final Map<LayerType, String> extractionPathMap = new EnumMap<>(LayerType.class);
 
     private Builder() {
       for (LayerType layerType : LayerType.values()) {
         layerFilesMap.put(layerType, new ArrayList<>());
+        extractionPathMap.put(layerType, "/");
       }
     }
 
-    public Builder setDependencyFiles(List<Path> dependencyFiles) {
+    public Builder setDependencyFiles(List<Path> dependencyFiles, String extractionPath) {
       layerFilesMap.put(LayerType.DEPENDENCIES, dependencyFiles);
+      extractionPathMap.put(LayerType.DEPENDENCIES, extractionPath);
       return this;
     }
 
-    public Builder setSnapshotDependencyFiles(List<Path> snapshotDependencyFiles) {
+    public Builder setSnapshotDependencyFiles(
+        List<Path> snapshotDependencyFiles, String extractionPath) {
       layerFilesMap.put(LayerType.SNAPSHOT_DEPENDENCIES, snapshotDependencyFiles);
+      extractionPathMap.put(LayerType.SNAPSHOT_DEPENDENCIES, extractionPath);
       return this;
     }
 
-    public Builder setResourceFiles(List<Path> resourceFiles) {
+    public Builder setResourceFiles(List<Path> resourceFiles, String extractionPath) {
       layerFilesMap.put(LayerType.RESOURCES, resourceFiles);
+      extractionPathMap.put(LayerType.RESOURCES, extractionPath);
       return this;
     }
 
-    public Builder setClassFiles(List<Path> classFiles) {
+    public Builder setClassFiles(List<Path> classFiles, String extractionPath) {
       layerFilesMap.put(LayerType.CLASSES, classFiles);
+      extractionPathMap.put(LayerType.CLASSES, extractionPath);
       return this;
     }
 
-    public Builder setExtraFiles(List<Path> extraFiles) {
+    public Builder setExtraFiles(List<Path> extraFiles, String extractionPath) {
       layerFilesMap.put(LayerType.EXTRA_FILES, extraFiles);
+      extractionPathMap.put(LayerType.EXTRA_FILES, extractionPath);
       return this;
     }
 
     // TODO: remove this and put files in WAR into the relevant layers (i.e., dependencies, snapshot
     // dependencies, resources, and classes layers).
-    public Builder setExplodedWarFiles(List<Path> explodedWarFiles) {
+    public Builder setExplodedWarFiles(List<Path> explodedWarFiles, String extractionPath) {
       layerFilesMap.put(LayerType.EXPLODED_WAR, explodedWarFiles);
+      extractionPathMap.put(LayerType.EXPLODED_WAR, extractionPath);
       return this;
     }
 
@@ -113,19 +118,26 @@ public class JavaLayerConfigurations {
       ImmutableMap.Builder<LayerType, LayerConfiguration> layerConfigurationsMap =
           ImmutableMap.builderWithExpectedSize(LayerType.values().length);
       for (LayerType layerType : LayerType.values()) {
+        String extractionPath = Preconditions.checkNotNull(extractionPathMap.get(layerType));
+        // Windows filenames cannot have "/", so this also blocks Windows-style path.
+        Preconditions.checkState(
+            extractionPath.startsWith("/"),
+            "extractionPath should be an absolute path in Unix-style: " + extractionPath);
+
         LayerConfiguration.Builder layerConfigurationBuilder =
             LayerConfiguration.builder().setName(layerType.getName());
 
         // Adds all the layer files recursively.
         List<Path> layerFiles = Preconditions.checkNotNull(layerFilesMap.get(layerType));
         for (Path layerFile : layerFiles) {
-          layerConfigurationBuilder.addEntryRecursive(
-              layerFile, layerType.getExtractionPath().resolve(layerFile.getFileName()));
+          Path pathInContainer = Paths.get(extractionPath).resolve(layerFile.getFileName());
+          layerConfigurationBuilder.addEntryRecursive(layerFile, pathInContainer);
         }
 
         layerConfigurationsMap.put(layerType, layerConfigurationBuilder.build());
       }
-      return new JavaLayerConfigurations(layerConfigurationsMap.build());
+      return new JavaLayerConfigurations(
+          layerConfigurationsMap.build(), ImmutableMap.copyOf(extractionPathMap));
     }
   }
 
@@ -133,11 +145,20 @@ public class JavaLayerConfigurations {
     return new Builder();
   }
 
+  /**
+   * The default app root in the image. For example, if this is set to {@code "/app"}, dependency
+   * JARs will be in {@code "/app/libs"}.
+   */
+  public static final String DEFAULT_APP_ROOT = "/app";
+
   private final ImmutableMap<LayerType, LayerConfiguration> layerConfigurationMap;
+  private final ImmutableMap<LayerType, String> extractionPathMap;
 
   private JavaLayerConfigurations(
-      ImmutableMap<LayerType, LayerConfiguration> layerConfigurationsMap) {
-    this.layerConfigurationMap = layerConfigurationsMap;
+      ImmutableMap<LayerType, LayerConfiguration> layerConfigurationMap,
+      ImmutableMap<LayerType, String> extractionPathMap) {
+    this.layerConfigurationMap = layerConfigurationMap;
+    this.extractionPathMap = extractionPathMap;
   }
 
   public ImmutableList<LayerConfiguration> getLayerConfigurations() {
@@ -170,7 +191,35 @@ public class JavaLayerConfigurations {
     return getLayerEntries(LayerType.EXPLODED_WAR);
   }
 
+  public String getDependencyExtractionPath() {
+    return getExtractionPath(LayerType.DEPENDENCIES);
+  }
+
+  public String getSnapshotDependencyExtractionPath() {
+    return getExtractionPath(LayerType.SNAPSHOT_DEPENDENCIES);
+  }
+
+  public String getResourceExtractionPath() {
+    return getExtractionPath(LayerType.RESOURCES);
+  }
+
+  public String getClassExtractionPath() {
+    return getExtractionPath(LayerType.CLASSES);
+  }
+
+  public String getExtraFilesExtractionPath() {
+    return getExtractionPath(LayerType.EXTRA_FILES);
+  }
+
+  public String getExplodedWarExtractionPath() {
+    return getExtractionPath(LayerType.EXPLODED_WAR);
+  }
+
   private ImmutableList<LayerEntry> getLayerEntries(LayerType layerType) {
     return Preconditions.checkNotNull(layerConfigurationMap.get(layerType)).getLayerEntries();
+  }
+
+  private String getExtractionPath(LayerType layerType) {
+    return Preconditions.checkNotNull(extractionPathMap.get(layerType));
   }
 }