+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/PluginConfigurationProcessor.java
@@ -39,6 +39,24 @@ import org.apache.maven.plugin.MojoExecutionException;
 class PluginConfigurationProcessor {
 
   /**
+   * Gets the value of the {@code <container><appRoot>} parameter. Throws {@link
+   * MojoExecutionException} if it is not an absolute path in Unix-style.
+   *
+   * @param jibPluginConfiguration the Jib plugin configuration
+   * @return the app root value
+   * @throws MojoExecutionException if the app root is not an absolute path in Unix-style
+   */
+  static String getAppRootChecked(JibPluginConfiguration jibPluginConfiguration)
+      throws MojoExecutionException {
+    String appRoot = jibPluginConfiguration.getAppRoot();
+    if (!ConfigurationPropertyValidator.isAbsoluteUnixPath(appRoot)) {
+      throw new MojoExecutionException(
+          "<container><appRoot> (" + appRoot + ") is not an absolute Unix-style path");
+    }
+    return appRoot;
+  }
+
+  /**
    * Sets up {@link BuildConfiguration} that is common among the image building goals. This includes
    * setting up the base image reference/authorization, container configuration, cache
    * configuration, and layer configuration.
@@ -108,7 +126,7 @@ class PluginConfigurationProcessor {
       String mainClass = projectProperties.getMainClass(jibPluginConfiguration);
       entrypoint =
           JavaEntrypointConstructor.makeDefaultEntrypoint(
-              jibPluginConfiguration.getJvmFlags(), mainClass);
+              jibPluginConfiguration.getAppRoot(), jibPluginConfiguration.getJvmFlags(), mainClass);
     } else if (jibPluginConfiguration.getMainClass() != null
         || !jibPluginConfiguration.getJvmFlags().isEmpty()) {
       logger.warn("<mainClass> and <jvmFlags> are ignored when <entrypoint> is specified");